test: move tests in `Security` file to specific categories
