+++ tests/LSP6KeyManager/index.ts
@@ -10,6 +10,7 @@ export * from "./SetPermissions/SetAllowedCalls.test";
 export * from "./SetPermissions/SetAllowedERC725YDataKeys.test";
 
 // Interactions
+export * from "./Interactions/InvalidExecutePayloads.test";
 export * from "./Interactions/PermissionCall.test";
 export * from "./Interactions/PermissionStaticCall.test";
 export * from "./Interactions/PermissionDelegateCall.test";
@@ -19,12 +20,12 @@ export * from "./Interactions/AllowedAddresses.test";
 export * from "./Interactions/AllowedFunctions.test";
 export * from "./Interactions/AllowedStandards.test";
 
+// Batch
+export * from "./Interactions/BatchExecute.test";
+
 // Relay
 export * from "./Relay/MultiChannelNonce.test";
 export * from "./Relay/ExecuteRelayCall.test";
-
-// Batch
-export * from "./Batch/BatchExecute.test";
 
 // Reentrancy
 export * from "./Reentrancy/Reentrancy.test";
@@ -34,8 +35,6 @@ export * from "./SetData/PermissionSetData.test";
 export * from "./SetData/AllowedERC725YDataKeys.test";
 
 // Others
-export * from "./OtherScenarios.test";
-export * from "./Security.test";
 export * from "./LSP6ControlledToken.test";
 
 // Internals (Unit Tests for internal functions)
